added Infrastructure layer and implemented ProductRepository

+++ ProductManager.Core/Services/ProductDeleterService.cs
@@ -1,13 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using AutoMapper;
-using ProductManager.Core.Domain.Entities;
+﻿using ProductManager.Core.Domain.Entities;
 using ProductManager.Core.Domain.RepositoryContracts;
 using ProductManager.Core.DTOs.ProductDTOs;
 using ProductManager.Core.ServiceContracts;
+using AutoMapper;
 
 namespace ProductManager.Core.Services
 {
@@ -15,6 +10,7 @@ namespace ProductManager.Core.Services
     {
         private readonly IProductRepository _productRepository;
         private readonly IMapper _mapper;
+
         public ProductDeleterService(IProductRepository productRepository, IMapper mapper)
         {
             _productRepository = productRepository;
@@ -23,31 +19,53 @@ namespace ProductManager.Core.Services
 
         public async Task<bool> DeleteProductAsync(Guid productId)
         {
-            var product = await _productRepository.GetProductByIdAsync(productId);
-            if (product == null) 
-                throw new ArgumentNullException(nameof(product));
-            var res = await _productRepository.DeleteProductAsync(product.ID);
-            return res;
+            if (productId == Guid.Empty)
+                throw new ArgumentException("آیدی محصول نمی‌تواند خالی باشد", nameof(productId));
+
+            var result = await _productRepository.DeleteProductAsync(productId);
+
+            if (!result)
+                throw new KeyNotFoundException($"محصول با آیدی {productId} یافت نشد");
+
+            return true;
         }
-        public async Task<bool> DeleteProductsAsync(List<ProductResponse?> products)
+
+        public async Task<bool> DeleteProductsAsync(List<ProductResponse> products)
         {
             if (products == null)
-                throw new ArgumentNullException(nameof(products));
+                throw new ArgumentNullException(nameof(products), "لیست محصولات نمی‌تواند خالی باشد");
 
-            var entities = _mapper.Map<List<Product>>(products);
+            if (!products.Any())
+                throw new ArgumentException("لیست محصولات نمی‌تواند خالی باشد", nameof(products));
 
-            var res = await _productRepository.DeleteProductsAsync(entities);
-            return res;
+            var validProducts = products.Where(p => p != null).ToList();
+
+            if (!validProducts.Any())
+                throw new ArgumentException("هیچ محصول معتبری برای حذف وجود ندارد", nameof(products));
+
+            var entities = _mapper.Map<List<Product>>(validProducts);
+            var result = await _productRepository.DeleteProductsAsync(entities);
+
+            return result;
         }
+
         public async Task<bool> DeleteProductsBeforeThan(DateTime date)
         {
-            var products = await _productRepository.GetAllProductsAsync();
-            if (products == null || products.Count == 0)
-                throw new ArgumentNullException(nameof(products));
+            if (date == DateTime.MinValue || date == DateTime.MaxValue)
+                throw new ArgumentException("تاریخ وارد شده معتبر نیست", nameof(date));
 
-            var productsToDelete = products.Where(p => p.Date < date).ToList();
-            var res = await _productRepository.DeleteProductsAsync(productsToDelete);
-            return res;
+            var allProducts = await _productRepository.GetAllProductsAsync();
+
+            if (allProducts == null || !allProducts.Any())
+                return false;
+
+            var productsToDelete = allProducts.Where(p => p.Date < date).ToList();
+
+            if (!productsToDelete.Any())
+                return false;
+
+            var result = await _productRepository.DeleteProductsAsync(productsToDelete);
+            return result;
         }
     }
 }
